Add the Kimberley, and make a region a registry entry rather than a page
The Dampier Peninsula is Sandover again. Ardyaloon, Djarindjin, Lombadina,
Beagle Bay and Bidyadanga are absent from ABS SAL_2021, and DAMPIER PENINSULA,
the locality above four of them, spans Broome and Derby-West Kimberley — so
resolving to it would still not name a council. Bidyadanga is the largest
remote Aboriginal community in Western Australia and the gazetteer has no entry
for it at all.

Broome is credited with $17.1M earned up the peninsula, across four
organisations checked one by one: Djarindjin Aboriginal Corporation, Ardyaloon
Incorporated, Bardi and Jawi Niimidiman, Ardyaloon Art & Culture. All four give
Broome as their ACNC town because that is where the post goes.

The wall is worse here than anywhere measured so far. Halls Creek has five
organisations with a council area and 100 in postcode 6770 without, 75 of them
community-controlled. Across the region 391 of 721 unplaced organisations are
community-controlled.

unplaced takes a list of postcodes now. The Kimberley's wall is spread across
five, and naming only the largest would have reported a fraction of it as the
whole. geo_resolution_gaps is summed across them rather than read as one row.

The report is one component now, driven entirely by the registry. Central
Australia and the Kimberley both render from it; Central Australia keeps its
map through a slot. This is the point of the registry work: the Far West Coast
entry spent a day describing the opposite of what the register said, and three
copies of the same prose is how that happens.

Verified by render, both pages 200. Central Australia still reports $93.6M
against the same five organisations, so the refactor moved nothing.
